requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pyodbc
pandas
openpyxl